Fixes multiple definition linking error for newer GCC/LD
The headers for the watchdog, rs485, and LED command components are
recursively include the overall RTD header in which they are also
included. This causes the various `CMD_*` constant structs to be
multiply defined in the component objects which is forbidden in the C
standards. Since the implementation for each of these command structs
appears only once, we declare them in the headers to be extern to avoid
multiple definition errors when linking `wdt.o`, `led.o`, and `rs485.o`
to `rtd.o`.